+++ cli/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add rust core as dependency for protocol and rs-bindings to ensure they will be rebuilt on changes in rust core.
 * Skip build, test, lint job if any of the build related jobs of their all dependencies failed.
 * Build related jobs for linting and testing can be skipped if possible.
+* Remove test and lint jobs from dependencies if the targets aren't included in the original targets.
 * Update dependencies
 
 # 0.3.0